+++ cloudauth-intl-20220809/alibabacloud_cloudauth_intl20220809/models/_ekyc_verify_v2response_body.py
@@ -71,15 +71,15 @@ class EkycVerifyV2ResponseBodyResult(DaraModel):
         sub_code: str = None,
         transaction_id: str = None,
     ):
-        # The face verification result information.
+        # The face result information.
         self.ext_face_info = ext_face_info
-        # The document recognition result. This parameter is returned only when the API response is successful.
+        # The identity document recognition result. Returned only when the API response is successful.
         self.ext_id_info = ext_id_info
-        # Indicates whether the verification is passed. Valid values: Y (passed) and N (not passed).
+        # Indicates whether the verification passed. A value of Y indicates passed, and a value of N indicates not passed.
         self.passed = passed
         # The sub-result code.
         self.sub_code = sub_code
-        # The unique identifier of the authentication request.
+        # The unique identifier of the verification request.
         self.transaction_id = transaction_id
 
     def validate(self):